+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/OfflinePlugin.kt
@@ -5,24 +5,31 @@ import androidx.annotation.VisibleForTesting
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.QueryChannelRequest
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
+import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.experimental.plugin.Plugin
 import io.getstream.chat.android.client.experimental.plugin.listeners.ChannelMarkReadListener
+import io.getstream.chat.android.client.experimental.plugin.listeners.EditMessageListener
 import io.getstream.chat.android.client.experimental.plugin.listeners.HideChannelListener
 import io.getstream.chat.android.client.experimental.plugin.listeners.QueryChannelListener
 import io.getstream.chat.android.client.experimental.plugin.listeners.QueryChannelsListener
 import io.getstream.chat.android.client.experimental.plugin.listeners.ThreadQueryListener
+import io.getstream.chat.android.client.extensions.cidToTypeAndId
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.livedata.utils.toLiveDataRetryPolicy
 import io.getstream.chat.android.offline.ChatDomainImpl
+import io.getstream.chat.android.offline.experimental.channel.logic.ChannelLogic
 import io.getstream.chat.android.offline.experimental.global.GlobalMutableState
 import io.getstream.chat.android.offline.experimental.global.GlobalState
 import io.getstream.chat.android.offline.experimental.plugin.logic.LogicRegistry
 import io.getstream.chat.android.offline.experimental.plugin.state.StateRegistry
+import io.getstream.chat.android.offline.extensions.isPermanent
+import java.util.Date
 
 /**
  * Implementation of [Plugin] that brings support for the offline feature.
@@ -39,16 +46,27 @@ public class OfflinePlugin(
     QueryChannelListener,
     ThreadQueryListener,
     ChannelMarkReadListener,
+    EditMessageListener,
     HideChannelListener {
 
     internal constructor() : this(Config())
 
+    /**
+     * [StateRegistry] which contains all states of this plugin.
+     */
     // TODO make it val and stateless when remove QueryChannelsMutableState::defaultChannelEventsHandler
     public lateinit var state: StateRegistry
         private set
+
+    /**
+     * [LogicRegistry] which contains all the logic to handle side effects.
+     */
     internal lateinit var logic: LogicRegistry
         private set
 
+    /**
+     * Global state of this plugin.
+     */
     // TODO: Move to StateRegistry when we remove ChatDomain.
     public val globalState: GlobalState = GlobalMutableState()
 
@@ -70,6 +88,42 @@ public class OfflinePlugin(
     internal fun initState(chatDomainImpl: ChatDomainImpl, chatClient: ChatClient) {
         state = StateRegistry(chatDomainImpl, chatClient)
         logic = LogicRegistry(state)
+    }
+
+    /**
+     * Method called when a message edit request happens. This method should be used to update messages locally and
+     * update the cache.
+     *
+     * @param message [Message].
+     */
+    override suspend fun onMessageEditRequest(message: Message) {
+        val (channelType, channelId) = message.cid.cidToTypeAndId()
+        val channelLogic = logic.channel(channelType, channelId)
+
+        val isOnline = globalState.isOnline()
+        val messagesToEdit = message.updateMessageOnlineState(isOnline).let(::listOf)
+
+        updateAndSaveMessages(messagesToEdit, channelLogic)
+    }
+
+    /**
+     * Method called when an edition in a message returns from the API.
+     *
+     * @param result the result of the API call.
+     */
+    override suspend fun onMessageEditResult(originalMessage: Message, result: Result<Message>) {
+        if (result.isSuccess) {
+            val message = result.data()
+            val channelLogic = channelLogicForMessage(message)
+            val messages = message.copy(syncStatus = SyncStatus.COMPLETED).let(::listOf)
+
+            updateAndSaveMessages(messages, channelLogic)
+        } else {
+            val channelLogic = channelLogicForMessage(originalMessage)
+            val failedMessage = originalMessage.updateFailedMessage(result.error()).let(::listOf)
+
+            updateAndSaveMessages(failedMessage, channelLogic)
+        }
     }
 
     override suspend fun onQueryChannelsRequest(request: QueryChannelsRequest): Unit =
@@ -122,6 +176,55 @@ public class OfflinePlugin(
     override suspend fun onChannelMarkReadPrecondition(channelType: String, channelId: String): Result<Unit> =
         logic.channel(channelType, channelId).onChannelMarkReadPrecondition(channelType, channelId)
 
+    /**
+     * Updates the messages locally and saves it at database.
+     *
+     * @param messages The list of messages to be updated in the SDK and to be saved in database.
+     * @param channelLogic [ChannelLogic].
+     */
+    private suspend fun updateAndSaveMessages(messages: List<Message>, channelLogic: ChannelLogic) {
+        channelLogic.upsertMessages(messages)
+        channelLogic.storeMessageLocally(messages)
+    }
+
+    /**
+     * Gets the channel logic for the channel of a message.
+     *
+     * @param message [Message].
+     */
+    private fun channelLogicForMessage(message: Message): ChannelLogic {
+        val (channelType, channelId) = message.cid.cidToTypeAndId()
+        return logic.channel(channelType, channelId)
+    }
+
+    /**
+     * Update the online state of a message.
+     *
+     * @param isOnline [Boolean].
+     */
+    private fun Message.updateMessageOnlineState(isOnline: Boolean): Message {
+        return this.copy(
+            syncStatus = if (isOnline) SyncStatus.IN_PROGRESS else SyncStatus.SYNC_NEEDED,
+            updatedLocallyAt = Date()
+        )
+    }
+
+    /**
+     * Updates a message that whose request (Edition/Delete/Reaction update...) has failed.
+     *
+     * @param chatError [ChatError].
+     */
+    private fun Message.updateFailedMessage(chatError: ChatError): Message {
+        return this.copy(
+            syncStatus = if (chatError.isPermanent()) {
+                SyncStatus.FAILED_PERMANENTLY
+            } else {
+                SyncStatus.SYNC_NEEDED
+            },
+            updatedLocallyAt = Date(),
+        )
+    }
+
     override suspend fun onHideChannelRequest(channelType: String, channelId: String, clearHistory: Boolean): Unit =
         logic.channel(channelType, channelId).onHideChannelRequest(channelType, channelId, clearHistory)
 
@@ -138,6 +241,9 @@ public class OfflinePlugin(
     }
 
     public companion object {
+        /**
+         * Name of this plugin module.
+         */
         public const val MODULE_NAME: String = "Offline"
     }
 }